Add apigateway to comunicate frontend and backend
