+++ Codigo_2B/GaleriaArte_V2.sln
@@ -15,6 +15,10 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "services", "services", "{0B
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UsuarioService", "src\backend\services\UsuarioService\UsuarioService.csproj", "{6D173F3E-DD02-4D1D-83BC-8D1CBFAB5240}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ObraService", "src\backend\services\ObraService\ObraService.csproj", "{BC06F601-504D-44CD-8004-2036FA4D0AC5}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ApiGateway", "src\backend\ApiGateway\ApiGateway.csproj", "{70C8F8AE-75D2-49A5-B6B6-1D5F07B5AC6F}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -32,6 +36,14 @@ Global
 		{6D173F3E-DD02-4D1D-83BC-8D1CBFAB5240}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{6D173F3E-DD02-4D1D-83BC-8D1CBFAB5240}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{6D173F3E-DD02-4D1D-83BC-8D1CBFAB5240}.Release|Any CPU.Build.0 = Release|Any CPU
+		{BC06F601-504D-44CD-8004-2036FA4D0AC5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{BC06F601-504D-44CD-8004-2036FA4D0AC5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{BC06F601-504D-44CD-8004-2036FA4D0AC5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{BC06F601-504D-44CD-8004-2036FA4D0AC5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{70C8F8AE-75D2-49A5-B6B6-1D5F07B5AC6F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{70C8F8AE-75D2-49A5-B6B6-1D5F07B5AC6F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{70C8F8AE-75D2-49A5-B6B6-1D5F07B5AC6F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{70C8F8AE-75D2-49A5-B6B6-1D5F07B5AC6F}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{492B82CC-8F54-4389-83A9-4002E308AE0D} = {587B63CD-B826-4EBE-842B-9EABA2EB1BE0}
@@ -39,5 +51,7 @@ Global
 		{511C1342-365B-4A81-955D-5505D57E450B} = {587B63CD-B826-4EBE-842B-9EABA2EB1BE0}
 		{0B98A2D1-20F1-447E-9081-2D6F1B667C26} = {511C1342-365B-4A81-955D-5505D57E450B}
 		{6D173F3E-DD02-4D1D-83BC-8D1CBFAB5240} = {0B98A2D1-20F1-447E-9081-2D6F1B667C26}
+		{BC06F601-504D-44CD-8004-2036FA4D0AC5} = {0B98A2D1-20F1-447E-9081-2D6F1B667C26}
+		{70C8F8AE-75D2-49A5-B6B6-1D5F07B5AC6F} = {511C1342-365B-4A81-955D-5505D57E450B}
 	EndGlobalSection
 EndGlobal
