+++ Codigo_2B/src/backend/ApiGateway/appsettings.json
@@ -1,0 +1,21 @@
+{
+  "ReverseProxy": {
+    "Routes": {
+      "usuarioTestRoute": {
+        "ClusterId": "usuarioCluster",
+        "Match": {
+          "Path": "/api/test/{**catch-all}"
+        }
+      }
+    },
+    "Clusters": {
+      "usuarioCluster": {
+        "Destinations": {
+          "usuarioService": {
+            "Address": "https://stunning-guacamole-xq67qrxqgv629qgr-5016.app.github.dev/"
+          }
+        }
+      }
+    }
+  }
+}
