+++ Codigo_2B/src/backend/services/UsuarioService/Controllers/TestController.cs
@@ -1,0 +1,15 @@
+using Microsoft.AspNetCore.Mvc;
+
+namespace UsuarioService.Controllers
+{
+    [ApiController]
+    [Route("api/[controller]")]
+    public class TestController : ControllerBase
+    {
+        [HttpGet("mensaje")]
+        public IActionResult ObtenerMensaje()
+        {
+            return Ok(new { mensaje = "¡Hola desde UsuarioService a través del API Gateway!" });
+        }
+    }
+}
